Add refactor code base task for sprint 2
Forgot to add this one to the sprint list. Essentially we
want to ensure that the project will be maintainable in the near
future and can be easily extended. To this end, we are refactoring
legacy code to uphold modern javascript best practices. Specifically
we will conform to Google's coding style conventions for javascript
from now on. Furthermore, project functionaility will be aggresively
partitioned into separate files representing modules.