Add simple zoom in/out functionaility
In the case where want to display large diagrams, users can focus
on area of interest at a time. Late down the road we need to add
proper zoom functionaility via scrollable #diagram-div. We can hide
the scrollbar and simply use the pointer position as a guide for the
horizontal displacement of the scroll.