Refactor output window/div
The word output seemed too vague so I replaced it with explorer output.
This may seem more verbose but I think it carries more semantic weight.
The textarea is replaced with a div because semantically a textarea
should be used for text input not read-only text. There is also the
unwanted side-effect of being able to resize the text area.

To complement this commit I also replaced raw innerHTML modification
with a familiar println()/print() family of functions provided by the
gui module. Gradually we rely on something similar to display text to
the user. Another good thing about this is that we can now support
a scroll history.

I'm also palying with the idea of styling the actual div but I haven't
found a good combination yet.