Encode verilog parameters within url
The javascript zip library we are using has a nice convinience function
for fetching zip files at remote endpoints. However, there isn't an
option to pass additional arguments such as a json dictionary.

As a side note, the server IP address is currently hardcoded so
we should try to find an alternative method to specify the IP
address.