Update memory block diagram to support 2 levels
For now since we only support up to two levels of cache, we should
show this structure in the diagram. In addition to this update,
the design of the diagram has changed to be simpler. Ideally in the
future we should support an arbitrary number of cache levels. Also
we still need to color this new diagram.