Show cache size updates on *_bit_width fields
Although I'm not sure how L1 cache size should be calculated, this
change now causes the memory diagram to reflect address_bit_width
and data_bit_width values chosen by the user. Sadly I just realized
that these fields are not within the memory tab and therefore
you won't see it update live in the current layout. There might
be some solutions to this but for now it is okay.