Fix bug where diagram won't switch on new stack
This bug is really annoying. Basically since we originally
tied the tab switch to the initial stack defined by the config
the switching will no longer work if the user creates a new stack
via dragging containers. I originally tried to use the container.show()
but that turned out to cause some strange bugs (containers
overlapping). It seems this is the most stable approach thus far.